TypeScript: Convert funcitons in @wordpress/wordcount package to TypeScript
