+++ packages/wordcount/README.md
@@ -29,9 +29,9 @@ const numberOfWords = count( 'Words to count', 'words', {} );
 
 _Parameters_
 
--   _text_ `string`: - The text being processed
--   _type_ `WPWordCountStrategy`: - The type of count. Accepts 'words', 'characters_excluding_spaces', or 'characters_including_spaces'.
--   _userSettings_ `WPWordCountUserSettings`: - Custom settings object.
+-   _text_ `string`: The text being processed
+-   _type_ `WPWordCountStrategy`: The type of count. Accepts 'words', 'characters_excluding_spaces', or 'characters_including_spaces'.
+-   _userSettings_ `WPWordCountUserSettings`: Custom settings object.
 
 _Returns_
